Act.4 - Movida lógica de ItemCount a ItemDetail
